Workaround for request body translation
