--- src/tests/contract/model/ApiGatewayEvents.ts
@@ -1,51 +1,0 @@
-import { randomUUID } from "crypto";
-
-export const eventRequest = {
-	body: "",
-	requestContext: {
-		requestId: randomUUID(),
-		accountId: "",
-		apiId: "",
-		authorizer: undefined,
-		protocol: "",
-		httpMethod: "",
-		identity: {
-			accessKey: "",
-			accountId: "",
-			apiKey: "",
-			apiKeyId: "",
-			caller: "",
-			clientCert: {
-				clientCertPem: "",
-				issuerDN: "",
-				serialNumber: "",
-				subjectDN: "",
-				validity: { notAfter: "", notBefore: "" },
-			},
-			cognitoAuthenticationProvider: "",
-			cognitoAuthenticationType: "",
-			cognitoIdentityId: "",
-			cognitoIdentityPoolId: "",
-			principalOrgId: "",
-			sourceIp: "",
-			user: "",
-			userAgent: "",
-			userArn: "",
-		},
-		path: "",
-		stage: "",
-		requestTimeEpoch: 0,
-		resourceId: "",
-		resourcePath: "",
-	},
-	headers: {},
-	multiValueHeaders: {},
-	httpMethod: "",
-	isBase64Encoded: false,
-	path: "",
-	pathParameters: null,
-	queryStringParameters: null,
-	multiValueQueryStringParameters: null,
-	stageVariables: null,
-	resource: "",
-};
