+++ src/tests/contract/utils/Constants.ts
@@ -1,0 +1,7 @@
+export class Constants {
+    
+    static readonly TOKEN_ENDPOINT = "/token";
+
+    static readonly USERINFO_ENDPOINT = "/userinfo";
+
+}
